requirements_2.txt
Package                       Version  
----------------------------- ---------
absl-py                       0.7.1    
astor                         0.8.0    
backports.functools-lru-cache 1.5      
backports.weakref             1.0.post1
certifi                       2019.6.16
chardet                       3.0.4    
conda                         4.3.16   
cycler                        0.10.0   
dask                          1.2.2    
enum34                        1.1.6    
fastparquet                   0.3.1    
feather-format                0.4.0    
funcsigs                      1.0.2    
futures                       3.3.0    
gast                          0.2.2    
google-pasta                  0.1.7    
grpcio                        1.22.0   
h5py                          2.9.0    
idna                          2.8      
Keras                         2.2.4    
Keras-Applications            1.0.8    
Keras-Preprocessing           1.1.0    
kiwisolver                    1.1.0    
llvmlite                      0.29.0   
Markdown                      3.1.1    
matplotlib                    2.2.4    
mock                          3.0.5    
numba                         0.45.0   
numpy                         1.16.4   
pandas                        0.24.2   
pip                           19.2.1   
progressbar2                  3.42.0   
protobuf                      3.9.0    
pyarrow                       0.14.1   
pycosat                       0.6.3    
pyparsing                     2.4.1.1  
python-dateutil               2.8.0    
python-utils                  2.3.0    
pytz                          2019.1   
PyYAML                        5.1.1    
requests                      2.22.0   
root-numpy                    4.8.0    
root-pandas                   0.7.0    
ruamel.ordereddict            0.4.14   
ruamel.yaml                   0.16.0   
ruamel.yaml.clib              0.1.0    
scikit-learn                  0.20.3   
scipy                         1.2.2    
seaborn                       0.9.0    
setuptools                    28.8.0   
singledispatch                3.4.0.3  
six                           1.12.0   
sklearn                       0.0      
subprocess32                  3.5.4    
tensorboard                   1.14.0   
tensorflow                    1.14.0   
tensorflow-estimator          1.14.0   
termcolor                     1.1.0    
thrift                        0.11.0   
urllib3                       1.25.3   
Werkzeug                      0.15.5   
wheel                         0.29.0   
wrapt                         1.11.2   
